CvR: + Assignments are now stored in bytes. + Minor tweaks to the SATSolver.

+++ apps/satin/sat/BreederSolver.java
@@ -49,7 +49,7 @@ public final class BreederSolver {
 	if( traceSolver ){
 	    System.err.println( "ls" + level + ": trying assignment var[" + var + "]=" + val );
 	}
-	ctx.assignment[var] = val?1:0;
+	ctx.assignment[var] = val?(byte) 1:(byte) 0;
 	int res;
 	if( val ){
 	    res = ctx.propagatePosAssignment( p, var, level );
