+++ apps/satin/sat/SeqSolver.java
@@ -42,7 +42,7 @@ public final class SeqSolver {
 	if( traceSolver ){
 	    System.err.println( "ls" + level + ": trying assignment var[" + var + "]=" + val );
 	}
-	ctx.assignment[var] = val?1:0;
+	ctx.assignment[var] = val?(byte) 1:(byte) 0;
 	int res;
 	if( val ){
 	    res = ctx.propagatePosAssignment( p, var, level );
